xy_market/pyproject.toml
[project]
name = "xy-market"
version = "0.1.0"
description = "Shared library for Agent Swarms ecosystem (MarketplaceBK, Buyer, Seller agents)"
authors = [
    { name = "XyAgent", email = "agent@xyber.inc" },
]
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.3",
    "x402==0.2.1",
    "cdp-sdk>=1.33.2",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.23.2",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "isort>=5.12.0",
    "aioresponses>=0.7.4",
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xy_market"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",
    "F",
    "I",
    "W",
    "C90",
    "N",
    "D",
    "UP",
    "B",
    "S",
]
ignore = [
    "D100",
    "D104",
    "D107",
    "D203",
    "D212",
    "S101",
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "D"]

[tool.mypy]
python_version = "3.12"
files = ["src/**/*.py"]
exclude = ["tests/", "__pycache__/"]
strict = true
ignore_missing_imports = true
show_error_codes = true
warn_return_any = false
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

